+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.19.0] - Unreleased
+
+### Changed
+
+- Added `--website-redirect` option.
+- Improved tests.
+- Refactored.
+- updated dependencies.
+
+### Fixed
+
+- Fixed `--tagging` option bug that does not work with `--server-side-copy` option.
+
+
 ## [1.18.0] - 2025-07-09
 
 ### Changed